src/requirements.txt
tqdm
tlp-classifier
rooted-tree-classifier
cyclepath-classifier
brt-classifier
psycopg2-binary
